initial and fix errors

+++ backend/src/main/java/ru/itmo/modeling/tatianagorbunova/wind/model/Measurement.java
@@ -6,14 +6,14 @@ import java.sql.Timestamp;
  * Created by tatiana.gorbunova on 13.01.2017.
  */
 public class Measurement {
-    private Double time;
+    private Long time;
     private Double speed;
 
-    public Double getTime() {
+    public Long getTime() {
         return time;
     }
 
-    public void setTime(Double time) {
+    public void setTime(Long time) {
         this.time = time;
     }
